MingW compatible build attempt

+++ BigBaseV2/src/util/math.hpp
@@ -23,9 +23,9 @@ namespace big::math
 
 		return Vector3
 		{
-			-sin(z) * num,
-			cos(z) * num,
-			sin(x)
+			static_cast<float>(-sin(z)) * num,
+			static_cast<float>(cos(z)) * num,
+			static_cast<float>(sin(x))
 		};
 	}
